(chore) rename localStorage-helpers->local-storage
