+++ frontend/src/state/auth.ts
@@ -1,4 +1,4 @@
-import { getUserFromLocalStorage } from "helpers/localStorage-helpers";
+import { getUserFromLocalStorage } from "helpers/local-storage";
 import { atom } from "recoil";
 
 export const currentUserState = atom<string | null>({
